Reducido uso de memoria

+++ proyecto/IA.cpp
@@ -18,20 +18,20 @@ class IA{
 		2- Casilla a la que mover la ficha
 	*/
 	/*AUN SOLO BUSCA CON PROFUNDIDAD 1!!!!*/
-	int[2] movimiento (Board t, int jugador){
+	byte[2] movimiento (Board t, byte jugador){
 		
-		int casillas[] = t-> getPositions(jugador);
+		byte casillas[] = t-> getPositions(jugador);
 		
-		int beneficio = 0;
+		byte beneficio = 0;
 		//pos1 tiene la posicion de la ficha antes de ser movida y pos2 después
-		int pos1 = -1;
-		int pos2 = -1;
+		byte pos1 = -1;
+		byte pos2 = -1;
 		//Por cada ficha..
-		for (int i=0; i<casillas.length; i++){
+		for (byte i=0; i<casillas.length; i++){
 			//Por cada movimiento
-			for (int j=0; j<4; j++){
+			for (byte j=0; j<4; j++){
 				if (posibleMov(t,i,j)){
-					int beneficio2 = this.beneficio(t, i, j, jugador);
+					byte beneficio2 = this.beneficio(t, i, j, jugador);
 					if (beneficio<beneficio2 || pos1 = -1){
 						beneficio = beneficio2;
 						pos1 = i;
@@ -46,16 +46,16 @@ class IA{
 	private:
 	
 	//Devuelve true si es posible hacer un movimiento y false en caso contrario
-	bool posibleMov (Board t, int pos, int mov){
+	bool posibleMov (Board t, byte pos, byte mov){
 		return !((pos<4 && mov==0) || (pos>20 && mov==2) ||
 			(pos==24 && mov==1) || (pos==0 && mov==3) ||
 			(t.getPlayer(pos)!=0));
 	}
 	
 	//Devuelve el tablero resultante tras efectuar un movimiento
-	Board nuevoBoard (Board t, int pos, int mov){
+	Board nuevoBoard (Board t, byte pos, byte mov){
 		if(posibleMov(pos, mov)){
-			int nuevapos = mueve(pos, mov);
+			byte nuevapos = mueve(pos, mov);
 			t.remove(pos);
 			t.set(pos);
 		}else{
@@ -65,15 +65,15 @@ class IA{
 	
 	//Devuelve el beneficio de un movimiento.
 	//Si en ese movimiento se pierde, el beneficio es -30
-	int beneficio (Board t, int pos, int mov, int jugador){
-		int toReturn = 0;
-		int newPos = t.newPosition(pos, mov);
+	byte beneficio (Board t, byte pos, byte mov, byte jugador){
+		byte toReturn = 0;
+		byte newPos = t.newPosition(pos, mov);
 		if(newPos == -1){
 			return newPos;
 		}else{
 			Board aux = copyBoard(t);
 			aux.move(pos, newPos);
-			for(int i=0; i<25; i++){
+			for(byte i=0; i<25; i++){
 				if(aux.getPlayer(i)==jugador && aux.blocked(i)){
 					toReturn-=1;
 					t.remove(i);
